refactor shared modules
