+++ beacon-chain/casper/BUILD.bazel
@@ -13,7 +13,8 @@ go_library(
         "//beacon-chain/params:go_default_library",
         "//beacon-chain/utils:go_default_library",
         "//proto/beacon/p2p/v1:go_default_library",
-        "//shared:go_default_library",
+        "//shared/bitutil:go_default_library",
+        "//shared/mathutil:go_default_library",
         "@com_github_ethereum_go_ethereum//common:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
     ],
