+++ shared/bitutil/BUILD.bazel
@@ -2,14 +2,14 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["service_registry.go"],
-    importpath = "github.com/prysmaticlabs/prysm/shared",
+    srcs = ["bit.go"],
+    importpath = "github.com/prysmaticlabs/prysm/shared/bitutil",
     visibility = ["//visibility:public"],
-    deps = ["@com_github_sirupsen_logrus//:go_default_library"],
+    deps = ["@com_github_steakknife_hamming//:go_default_library"],
 )
 
 go_test(
     name = "go_default_test",
-    srcs = ["service_registry_test.go"],
+    srcs = ["bit_test.go"],
     embed = [":go_default_library"],
 )
