+++ shared/mathutil/math_helper_test.go
@@ -1,4 +1,4 @@
-package shared
+package mathutil
 
 import (
 	"testing"
