+++ shared/shardutil/BUILD.bazel
@@ -2,14 +2,14 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["service_registry.go"],
-    importpath = "github.com/prysmaticlabs/prysm/shared",
+    srcs = ["marshal.go"],
+    importpath = "github.com/prysmaticlabs/prysm/shared/shardutil",
     visibility = ["//visibility:public"],
-    deps = ["@com_github_sirupsen_logrus//:go_default_library"],
+    deps = ["@com_github_ethereum_go_ethereum//rlp:go_default_library"],
 )
 
 go_test(
     name = "go_default_test",
-    srcs = ["service_registry_test.go"],
+    srcs = ["marshal_test.go"],
     embed = [":go_default_library"],
 )
